Add possibility to run Miniterm on serial port of board
This is handy for debugging. It allows you to access serial console
during tests and that way developer can access directly state it
requires.
The way how it works is that developer is going to call:
board.serial_miniterm()
in appropriate location.

To make this work pytest have to be run without terminal redirection.
This is as easy as appending -s to pytest call.