nsfarm/lxd: add ability to pass character devices
There is change that adds device exclusivity. Only for few devices we
need exclusive passtrough to container. In most cases it is enough to
pass only access to that device.

Another change is that there is possibility to specify parameter for
images. At the moment this allows only passtrough of Unix character
devices but that can be expanded in the future.