nsfarm/mterm: add micro terminal
The micro terminal is minimal connection between terminal and socket.
This way it can be hooked up to any socket in nsfarm and thus allowing
the direct access for CLI interfaces. The primary use is to access
board's serial console.

This replaces serial_miniterm with simpler and more generic implementation.