tests: remove serial and wan markers
The only possible execution right now is with serial and wan present.
In reality we need lan1 as well but we have to get around that because
CPU only Mox.
This simply removes parameters and marks for appropriate fixtures.
The effect is just less marks as well as better tests reports as they no
longer report None parameters.