selftests/lxd: add simple ping test for LXD container
The current tests check if appropriate profiles exists but not if they
actually work. The nsfarm-root has to work when container at least
starts. The problem is with nsfarm-internet. It is required to test if
it is possible to access the Internet or not with that profile. That is
exactly what this test does.