nsfarm/lxd: include files in hash calculations
There was for some reason condition that only directories are listed
when image hash is generated. In reality the code was prepared to handle
any path and should include all paths. This removes this invalid
condition and makes it so all files provided for image are included in
hash calculation.

This in effect makes it so new image build is triggered on any file
modification.