nsfarm/lxd: improve warning about image bootstrap
The message was being printed almost right on prepare method enter. The
problem is that it was called before parent image prepare method was
called. In effect this generated messages in reverse order. It also made
it pretty much impossible to identify which container exactly failed to
prepare as warnings for all images were printed even before any of them
started actually preparing. This moves it after parent prepare method
call and thus makes it so prepare messages are printed in order and
right before actual work is being performed.

This also gives us the option to include full container name we use for
bootstrapping.