nsfarm/mterm: suppress logging
It is common that tests are run with logs printed to terminal. It helps
developer to see what is actually happening.
The problem is that log printing interferes with terminal output. It
should be safe to set level of root logger such it prints no messages
for time of mterm execution.