tests/test_bootup: check if services we want are enabled
This checks not only if services we want are running but also if they
are enabled.
The real reason for this is not to cover issue of disabled but running
service as that is highly unlikely. This rather covers services that
spawn actually no process but still we have to check if they are enabled
and thus executed in some manner. There is a lot of such services.