nsfarm: allow and disable input read for serial console
This reason for this is that we use at some point serial console only to
get logs. We do not process it further. There is no reader for serial
console at most of time.
This change allows disable of serial console input and serial console
logs are thus only logged without further propagation.