tests/mark: add set of simple marks based on board mark
The idea behind this is that we have more descriptive names of marks and
thus it should be easier to understand why some tests are marked with
given combination of boards as they are.
At the same time it should be easier to go trough them at one location
once we add new boards. The new board would have to be selectively added
to appropriate marks and all tests using these marks would be
automatically handled. Of course that does not solve it fully but still
it reduces the burden on introducing new boards.