tests: replace basic_isp fixture with two fixtures
This way we can have easy access to ISP container as well as to WAN
configuration.
This is required as we have to know WAN IPv4 address in some generic
way. We could define constant but this way it is prepared for standard
ISP being based rather on DHCP in the future.