imgs/base-alpine: improve bootstrap script
This changes the order of setup. The original pretty much relied on
configuration change happening before service actually started and
system booted. This of course can't be ensured and is pretty fragile.

In general we can change file access or any file content without waiting
for system boot but once we want to communicate with services or to
access the Internet we need to wait for system to actually boot.

The clean effect here is the need to reload networking service once we
modify interfaces after boot. This should have always been there.