test_bootup: check not only for config for disabled network on boot
This adds additional test that tries if router really does not try to
receive address from DHCP even if not configured. This is even more
powerful test but thanks to need to blindly wait it can take some
considerable amount of time and thus does not replace test_no_wan but
complements it.