tests: start using deploy mark
The tests added in df0bf8d057735d316fc35f36928b2875b39e3f87 made it
necessary to start using deploy mark. The deploy mark is used on tests
needed to pass for deploy to proceed. These should be minimal necessary
tests and should be fast enough.
The reasoning why we need it here is because in most cases it should be
enough the check in the test_no_wan that runs fast. The additional
slower but more general test TestNoInternetAccess should cover only
corner case and thus is not essentially required as a blocker for
deploy.