tests: move target specification to top level
This allows to run selftests not on all available targets but only on
specified one. This way we do not pollute the results with test of any
other target.

This also improves marks initialization. Marks with limited effect
(those are lan1 and lan2) are defined in runtime. This makes their
definition and effect at the same level and thus we should be able to
better manage them.