selftests/target: improve check for interface being up
This changes check from carrier being up to just interface being up.
That is actually the correct check we want to do. The carrier might be
down if we are connected directly to board (not trough switch) and thus
this test might fail even if everything is all right. In reality we want
to check that interface is set to be up and thus is going to be up once
carrier is up. This information is encoded in flags as lowest bit (this
was discovered by checking flag value rather not from documentation). So
this test now only checks if this bit is set or not.