conftest: make pytest-html optional
We do not need pytest-html for normal functionality. It is just nice
plugin we can use to generate human readable report to go alongside with
pytest native xml and log output.
We also support only pytest-html with minimal version 2.1.0 and thus all
this is included only if pytest-html is at least of that version.

The plugin is removed by this from requirements.txt as this file lists
necessary Python packages and this is now only optional.