tests: use ping instead of sleep to check for network readiness
There is an issue that although software reports interface being up a
few initial frames won't pass. The original "fix" for it was to just
wait a moment but that brakes depending on hardware we are running tests
on. Although it is not ideal we instead use ping. Pinging our gateway
that is our ISP container should be enough to make sure that connection
is established.
At the same time this is not in any way required for DHCP. The reason is
because to get IP address we have to communicate with ISP and thus
connection is established for sure. This means that any sleep is nor any
other wait is required there.