imgs/base-alpine: fix mount of /dev/shm
The Alpine Linux creates /dev/shm using init script devfs but that is
not run in container. It won't even run correctly in container. That
makes /dev/shm unavailable but we need it in some cases and thus this is
hacky script that adds just very small init script that creates it.