imgs/selenium: add new image
This is image that can be used to test web interface.

It is based on client container but we do not want to put this as part
of it. It generates pretty big image and in most tests it is not
necessary to start web drivers.

The container itself provides driver for Chrome (Chromium) and Firefox
browsers.
There is also driver for WebKit but it does not work, it crashes right
now so it is there just for the future once something is fixed in
upstream or we found out what we are doing wrong.