tests/test_bootup: do not match exactly
The issue here is that some applications such as sshd are modifying its
own arguments to store readable info in there. This breaks check for
processes but they keep its name intact and so we can match that.

The -x is replaced with -a to show process info in logs so we can
identify if we match something that is not suppose to be matched.