nsfarm/web: add Selenium container class
This implements dedicated Selenium container class as an extension to
LXD Container. This class provides easy access to Selenium drivers for
selected browser.
The special hack in Selenium extended Container class is that it is able
to run vncviewer on its start. This is controlled by class variable that
can be set for nsfasrm run. This way developer can select to see what is
happening on virtual desktop browsers are spawned to.

With all this there is also need to store some screenshots as part of
failed or even standard execution of tests. The screenshot fixture was
created for this purpose. It hides some ugliness of screenshot saving.

Minor but notable tweak here is addition of shell property to
container. The idea is to prevent spawning of multiple one-time use
shells we do for wait commands. The regular usage for testing is
discouraged.