nsfarm/lxd: fix image hash calculation
The issue here is that we should not use full path to file never to
calculate hash as different location of NSFarm project result in
different hashes and thus we do not share them at all.
This makes one instance where we included full path just relative to
images directory. This way the path should be contant no matter where
the project is located.

This also does small tweak in form of removal of unnecessary
comprehension error. We can convert directly to list without creating
generator.