nsfarm/board: do not use target name in logging
We do not run tests on multiple boards at the same time and it is
technically undesirable in terms of how current tests are setup. That
makes inclusion of target name just a duplicate value as target is
stated in meta-data for tests and thus directly visible at the start
test report.