tests: decrease scope of fixture to just package
The scope of session reaches not only tests but also selftests and in
general we want to keep those two separate thus these fixtures should be
limited by package they are defined in.

In some cases this is rather extension because for some reason for
example client_board was defined with module scope. The effect of that
was that client was been reconnected for every module. Making that for
whole package saves a little bit of time. The original idea was to make
sure that we are connected but if ever test plays with local network it
on teardown it has to restore original settings and thus it should make
original connection available again.