tests/conftest: allow a unique shell for every test
This should increase tests resilience as we can now spawn new shell
instance for every test and thus taint of shell from one test can't
influence others.

This of course needs changes in existing tests and that is not part of
this code. The update of existing code and removal of client_board
fixture in favor of board_access should be done in the future. For now
client_board is documented as obsolete and not suggested for the new
code.

The disadvantage of this change is that every fixture that wants to do a
simple call (such as uci or cat) has to create its own connection and
that can get spammy pretty fast. The solution for this is the new
fixture fixture_board_access. This provides one shell instance. In
reality it provides shell instance that can reconnect. This is due to
issues when client container would change its IP address (for example
when we change LAN addressing).