nsfarm/lxd/container: format using black and allow name specification
This allows you to specify name for container. This name is used only in
logs and if not specified the image name is used. This should
primarily be used when you spawn multiple containers from same image
side by side. Then specifying name allows easy recognition of different
containers in the logs.