tests/conftest: change scope of top level fixtures to session
This was causing issue when tests from multiple packages were run in one
session as serial console is not correctly closed on cleanup of board
fixture.

These should be session anyway as they do not cause any conflict with
anything else (compare to boot_serial that boots and reboots the
router). There is no reason why we would want the new instances of these
objects, rather the opposite.