selftests/cli: use lorem-text instead of static lorem ipsum
This replaces static strings with generated ones from lorem-text
library.

To make tests reproducible we have to se seed for random. This is pretty
common so this adds context function for it to toolbox.

Toolbox was labeled as common tasks run on router. Now this changes it
to just generic common toolbox and splits it to three submodules. There
is module for Alpine Linux as well as for OpenWrt. Then there is one
additional module for utilities needed in tests implementation.