selftests: add test for bootstrap of all images
The common development workflow is to run selftest after changes. Such
significant change is for example update of base container. The issue is
that not every container is used in selftests and thus not every issue
is discovered early before tests on board are executed. It is possible
to run `./tool.sh lxd bootstrap -a` to check but that is easy to forget.
On the other hand this forces bootstrap of all images even if they are
not used in tests and thus possibly unnecessary work. That is a good
trade off for making sure that all needed images can be prepared.