tests/reforis: fix inaccessible router
This changes configuration for client container from static IP to DHCP.
The reason is because we spawn multiple clients in parallel (especially
with selenium) and having them all on same static IP does not work. At
the same time we want to have the primary client container on static IP
so when we test DHCP we won't break our primary connection to the
router.
Thus this converts the client-dhcp to client-static and makes client a
DHCP.

This also cleanups the client-static (previously client-dhcp) bootstrap
script. There is no need to install software that is already installed
in client.