tests: fix not working Internet access after WAN network tests
The WAN tests configure and reload network settings to test various
Internet access methods. The issue here is simply that after that is
done and configuration is reverted it takes some time before the
Internet access is reestablished from DHCP. Thus this twists existing
code to allow call to generic function for waiting for uplink in an
appropriate moment.