tests,selftests: use yield in fixtures instead of return
This requires some explanation specially because I was unable to locate
the root cause of the weirdness. The initial find was in selftests. The
selftests were passing on their own but once they were included in one
test run with tests the test cli.test_shell.test_txt[OpenWrt] started
failing but in setup of the fixture shell. It seemed that the new shell
session was forked but not exactly correctly connected. The shell in the
end was unable to read anything. At the same time there was output from
previous pexpect instances (thanks to logging). So it seems like somehow
the pexpect instances were messed up. When I added log message before
return it turned out that not on every test it was called and log (where
there should have been busybox banner) confirmed that. The "fix" by
changing return to yield was discovered by random as I switched it
around to add also log message for teardown. That suddenly started
working and issue disappeared. I tested a multiple combination and I
can't found out what triggers this behavior. I really think that this is
a bug in Pytest. It for sure handles fixtures that are just functions
and not generators slightly differently. Just to be sure I switched all
returns in fixtures to yields.