nsfarm/cli: get rid of exit_code and replace it with check
The original idea of exit_code was to allow flexibility of handling exit
code. The true is that most of our uses were to set it to None to
disable the default handler to get exit code. This use case now covers
check attribute instead. Any other use can be covered by getting exit
code from run method and doing anything with it later on.

The check attribute was selected to be consistent with standard
subprocess.run.