nsfarm/target: allow boot command with any ISP and client container
This adds a way to select ISP and client container when booting board
using tool. The defaults are used as expected but any other image can be
specified instead. To disable the empty argument can be provided
instead.

To implement this in more generic way this also adds a helper method to
image to identify if it is child of some other image. This is after that
used in dedicated function in nsfarm.setup.uplink that provides a
selection of best uplink setup for given ISP container.