nsfarm/web/reforis/guide: set default timeout for wait to one minute
The original 10 seconds were just too short time considering that
sometimes we have to reload newtork configuration before we proceede to
the next step in the guide. The tests that do that such as interfaces
were failing because of that. The one minit is reasonable time we can
give it before we consider it a failed state.